+++ src/components/playlist/SongList/SongMenu.tsx
@@ -3,8 +3,8 @@ import { Menu } from 'react-native-paper';
 import { useTranslation } from 'react-i18next';
 
 import { useNoxSetting } from '@stores/useApp';
-import { CopiedPlaylistMenuItem } from '../buttons/CopiedPlaylistButton';
-import { RenameSongMenuItem } from '../buttons/RenameSongButton';
+import { CopiedPlaylistMenuItem } from '../../buttons/CopiedPlaylistButton';
+import { RenameSongMenuItem } from '../../buttons/RenameSongButton';
 import useSongOperations from '@hooks/useSongOperations';
 enum ICONS {
   SEND_TO = 'playlist-plus',
